pyproject.toml
[project]
name = "lemarche"
version = "1.0"
description = "Marché de l'inclusion"
authors = []
requires-python = ">=3.13,<3.14"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "boto3<1.36.0",
    "brevo-python>=1.0",
    "crawlerdetect>=0.1,<0.2",
    "django>=5.2,<6",
    "django-admin-list-filters>=1.2",
    "django-allauth>=65.0",
    "django-anymail>=12.0,<13",
    "django-better-admin-arrayfield>=1.4",
    "django-ckeditor>=6.0",
    "django-compressor>=4.4",
    "django-cors-headers>=3.0,<4",
    "django-datadog-logger>=0.7",
    "django-dsfr==1.4.3",
    "django-environ>=0.9,<0.10",
    "django-extensions>=3.0,<4",
    "django-filter>=23.0,<24",
    "django-formtools>=2.0",
    "django-htmx>=1.0",
    "django-libsass>=0.9",
    "django-phonenumber-field[phonenumbers]>=7.3.0,<8",
    "djangorestframework>=3.0",
    "django-sesame>=3.0",
    "django-simple-history>=3.7.0",
    "django-storages>=1.0",
    "drf-spectacular>=0.26,<0.27",
    "huey>=2.0",
    "psycopg[binary]>3",
    "python-stdnum>=1.20",
    "redis>=5.0,<6",
    "requests>=2.0",
    "sentry-sdk>=2.0",
    "shortuuid>=1.0",
    "xlwt>=1.0",
    "wagtail>=6.1,<7",
    "wagtail-localize>=1.9,<2",
    "wagtail-markdown>=0.12.1,<0.13",
    "wagtail-modeladmin>=2.0.0,<3",
    "wagtailmenus>=4.0,<5",
    "unidecode>=1.3.8",
]

[dependency-groups]
dev = [
    "coverage>=7.0,<8",
    "django-debug-toolbar>=5.0,<6",
    "factory-boy>=3.0,<4",
    "freezegun>=1.5,<2",
    "ipdb>=0.13,<0.14",
    "pre-commit>=4.0,<5",
    "pyproject-flake8>=7.0,<8",
    "ruff",
    "selenium>=4.0,<5",
    "tblib>=3.0,<4",
]

[tool.ruff]
line-length = 119

[tool.ruff.lint]
# see prefixes in https://beta.ruff.rs/docs/rules/
select = [
    "F",  # pyflakes
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "I",  # isort
    "UP",  # pyupgrade
    "TID",  # flake8-tidy-imports
]

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["itou"]
lines-after-imports = 2

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "lemarche.settings.test"
